Fixed prototype mismatch in screen.h.  Change type of count in get_com_char to int16_t.  Combined set_kp_keys and set_cur_keys into set_keys, defined macros to replace old calls.

+++ cmdtok.c
@@ -44,7 +44,7 @@ static int16_t get_com_char(const int8_t flags)
 	if (flags & BUF_ONLY)
 		return(GCC_NULL);
 
-	int count;
+	int16_t count;
 
 	for (;;) {
 		fd_set in_fdset;
